+++ Sources/Qualtive/Controllers/LoggingController.swift
@@ -2,18 +2,18 @@ import Foundation
 import os
 
 /// Logs library diagnostics such as hints about newer SDK versions.
-public protocol LoggingControllerType: Sendable {
+protocol LoggingControllerType: Sendable {
 
   /// Logs that the client may need updating when an unknown API value is encountered.
   func logHintNewVersion()
 }
 
 /// Default logging using the system unified logging facility.
-public struct LoggingController: LoggingControllerType {
+struct LoggingController: LoggingControllerType {
 
-  public init() {}
+  init() {}
 
-  public func logHintNewVersion() {
+  func logHintNewVersion() {
     let text: StaticString =
       "There may be a new version of the Qualtive Client Library - Swift. Please update to get the latest features and fixes."
     os_log(text, log: .qualtive)